Check presence of required attributes in `Facade`
Two things are happening here. First, `Facade`'s `__init__` now has the
optional keyword argument `_facade_requires_` telling the constructor
which attributes to check for. In order to get this working, `Facade`
had to be made a subclass of `oemof`'s `Node`, a this ensures that
`Facade`'s constructor is called before `Node`'s. This in turn is
necessary because `Node` can't relay it's constructor arguments to a
superclass as that may be `object`, which doesn't accept arguments in
it's constructor.