Fix error message
Since attribute errors look a bit different than key errors, generating
the error message had to be revised to pull the attribute name out of
both of them. Fortunately it's as easy as pulling the last string
wrapped in single quotes out of the error message.
Positive side effect: I could remove these single quotes so that the
attribute name no longer looks like a string in the error message.
While at it, I also changed the message to output the label's `repr`, so
that the label can now correctly be identified as a string if it is one.