Fix checking for required attributes
Oh, the joys of lazy languages. I totally forgot that Python always
evaluates function arguments, so in:

```python
setattr(self, r, kwargs.get(r, getattr(self, r)))
```

the inner `getattr` call would always be evaluated which leads to a
false error if the attribute is in `kwargs` but not yet set on `self`.
This is now fixed.
As putting that long `if` statement into a `lambda` wouldn't really make
sense, I removed the exercise comment too. It also had a bug because I
was to lazy to copy the code line above completely and mistyped it
instead. D'Oh.
Uplifting point: with the new code I no longer need to parse the
parameter name out of the error message. Yay for less code doing the
same thing. (Technically this wasn't necessary before too, as `r` was
probably still in scope during the `except` block, but, whatever. :)